fix: more generic BlockCodec, remove encoder & decoder props
This is what we probably should have done to solve the conflict in cea9063c
because we ended up with a `BlockCodec` that has too many properties and is
therefore not very useful.

This makes `BlockCodec` just an implementer of both `BlockEncoder` and
`BlockDecoder` but `codec()` returns you that, plus `encoder` and `decoder`
properties so you have all the things and can consume them however you like.

Ref: https://github.com/ipld/js-dag-cbor/pull/18#issuecomment-814937133